refactor manual registration to use function
