+++ src/CMakeLists.txt
@@ -1,6 +1,6 @@
 cmake_path(SET SHARED_INCLUDE ${CMAKE_CURRENT_LIST_DIR})
 
-target_sources(deluge PUBLIC main.c resetprg.c fault_handler.c c_lib_alternatives.S sys_stubs.c)
+target_sources(deluge PUBLIC main.c resetprg.c fault_handler.c c_lib_alternatives.S sys_stubs.c timers_interrupts.c)
 
 add_subdirectory(deluge)
 add_subdirectory(RZA1)
